+++ src/backend/vulkan/BufferUploader.cpp
@@ -14,9 +14,9 @@
 
 #include "backend/vulkan/BufferUploader.h"
 
+#include "backend/vulkan/DeviceVk.h"
 #include "backend/vulkan/FencedDeleter.h"
 #include "backend/vulkan/MemoryAllocator.h"
-#include "backend/vulkan/VulkanBackend.h"
 
 #include <cstring>
