+++ src/backend/vulkan/ComputePipelineVk.cpp
@@ -14,10 +14,10 @@
 
 #include "backend/vulkan/ComputePipelineVk.h"
 
+#include "backend/vulkan/DeviceVk.h"
 #include "backend/vulkan/FencedDeleter.h"
 #include "backend/vulkan/PipelineLayoutVk.h"
 #include "backend/vulkan/ShaderModuleVk.h"
-#include "backend/vulkan/VulkanBackend.h"
 
 namespace backend { namespace vulkan {
