+++ src/backend/vulkan/QueueVk.cpp
@@ -15,7 +15,7 @@
 #include "backend/vulkan/QueueVk.h"
 
 #include "backend/vulkan/CommandBufferVk.h"
-#include "backend/vulkan/VulkanBackend.h"
+#include "backend/vulkan/DeviceVk.h"
 
 namespace backend { namespace vulkan {
