+++ src/backend/vulkan/RenderPipelineVk.cpp
@@ -16,13 +16,13 @@
 
 #include "backend/vulkan/BlendStateVk.h"
 #include "backend/vulkan/DepthStencilStateVk.h"
+#include "backend/vulkan/DeviceVk.h"
 #include "backend/vulkan/FencedDeleter.h"
 #include "backend/vulkan/InputStateVk.h"
 #include "backend/vulkan/PipelineLayoutVk.h"
 #include "backend/vulkan/RenderPassCache.h"
 #include "backend/vulkan/RenderPassDescriptorVk.h"
 #include "backend/vulkan/ShaderModuleVk.h"
-#include "backend/vulkan/VulkanBackend.h"
 
 namespace backend { namespace vulkan {
