+++ src/backend/vulkan/SamplerVk.cpp
@@ -14,8 +14,8 @@
 
 #include "backend/vulkan/SamplerVk.h"
 
+#include "backend/vulkan/DeviceVk.h"
 #include "backend/vulkan/FencedDeleter.h"
-#include "backend/vulkan/VulkanBackend.h"
 
 namespace backend { namespace vulkan {
